"Moved Constants to util package, added 2 topic builders, changed registerPlayer "key" to email instead.
