--- src/test/java/com/work/javaproducer/JavaProducerApplicationTests.java
@@ -1,14 +1,0 @@
-package com.work.javaproducer;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-// Todo Some tests here too?
-@SpringBootTest
-class JavaProducerApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
